Name-based seed...
Easier to retrieve/search for without a bookmark...
For example, when coming back to a longer playthrough.
Gender-neutral names so character sex is irrelevant.
(Neutral name list also contributed to `dariusk/corpora`)!
List of ~700 names may fall short of all RNG possibilities...
but it's more than enough for ~100 party combinations, which is the main factor.
(Or just right for ~700 progressive combinations leading up to final party).